+++ src/App/VRChatContentPublisher.App/ViewModels/MainWindowViewModel.cs
@@ -1,7 +1,9 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using VRChatContentPublisher.App.Localization;
 using VRChatContentPublisher.App.Services;
 using VRChatContentPublisher.App.Services.Dialog;
+using VRChatContentPublisher.App.Services.Notification;
 using VRChatContentPublisher.App.ViewModels.Pages;
 using VRChatContentPublisher.Core.Settings;
 using VRChatContentPublisher.Core.Settings.Models;
@@ -11,6 +13,8 @@ namespace VRChatContentPublisher.App.ViewModels;
 public partial class MainWindowViewModel : ViewModelBase, INavigationHost, IAppWindow
 {
     private readonly DialogService _dialogService;
+    private readonly IWritableOptions<AppSettings> _appSettings;
+    private readonly DesktopNotificationService _desktopNotificationService;
     [ObservableProperty] public partial PageViewModelBase? CurrentPage { get; private set; }
 
     [ObservableProperty] public partial bool Pinned { get; private set; }
@@ -24,10 +28,13 @@ public partial class MainWindowViewModel : ViewModelBase, INavigationHost, IAppW
         NavigationService navigationService,
         DialogService dialogService,
         AppWindowService appWindowService,
-        IWritableOptions<AppSettings> appSettings
+        IWritableOptions<AppSettings> appSettings,
+        DesktopNotificationService desktopNotificationService
     )
     {
         _dialogService = dialogService;
+        _appSettings = appSettings;
+        _desktopNotificationService = desktopNotificationService;
 
         SetBorderless(appSettings.Value.UseBorderlessWindow);
         DialogHostId = dialogService.DialogHostId;
@@ -60,5 +67,17 @@ public partial class MainWindowViewModel : ViewModelBase, INavigationHost, IAppW
     public void Activate()
     {
         RequestActivate?.Invoke(this, EventArgs.Empty);
+    }
+
+    public void NotifyWindowHide()
+    {
+        if (_appSettings.Value.DismissFirstTimeHideWindowTip)
+            return;
+
+        _appSettings.Update(settings => settings.DismissFirstTimeHideWindowTip = true);
+        _ = _desktopNotificationService.SendNotificationAsync(
+            LangKeys.Notifications_First_Time_Hide_Window_Tip_Title,
+            LangKeys.Notifications_First_Time_Hide_Window_Tip_Message
+        ).AsTask();
     }
 }