+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Will show tooltip for task creation time. [`#403`](https://github.com/project-vrcz/content-publisher/pull/403)
 - Add `ApplicationLifetimeSession` and `app.lifetime_session_id` to logging and tracing to identify each session of app.
 - Allow choice app language in onboarding. [`#399`](https://github.com/project-vrcz/content-publisher/pull/399)
+- Send desktop notification to tell user app hide to system tray when window hide at first time. [`#472`](https://github.com/project-vrcz/content-publisher/pull/472)
 - Preview changelog rendered from markdown in update dialog. [`#435`](https://github.com/project-vrcz/content-publisher/pull/435)
 - Show restart onboarding in-app notification when skip onboarding on first launch. [`#396`](https://github.com/project-vrcz/content-publisher/pull/396)
 - Add request background public ip check run button in debug settings. [`#456`](https://github.com/project-vrcz/content-publisher/pull/456)